Update IntegrationAndTesting.md
There is no link to the Migration guide when you say "If you were using surveys prior to version 1.5.0 of the Apptentive Android SDK, see this Migration Guide for
instructions."

Do you think it would be helpful to include more pictures of the functionality you are describing? For example, showing what a button linked to message center might look like, or is that overkill?